[Automated Tests] Refactor functions
